Refactor out Debrief action wrapper
